+++ plugins/kubernetes-react/src/index.ts
@@ -1,3 +1,4 @@
 export * from './components';
 export * from './hooks';
 export * from './lib';
+export { generateUID } from './utils/generateUID';